Corrected error regarding errormessage, tweaked importing of PropTypes to be more efficient. Moved components to components dir
