+++ src/components/QuestionPage.jsx
@@ -1,15 +1,26 @@
 import React from 'react';
+import {
+    array,
+    object,
+    arrayOf,
+    string,
+    func,
+    bool,
+    number,
+} from 'prop-types'
 import Grid from '@material-ui/core/Grid';
-import { Typography } from '@material-ui/core';
+import {
+    Typography,
+} from '@material-ui/core';
 import FreeText from '../containers/FreeText'
-import PropTypes, { array } from 'prop-types'
 import BigButton from '../components/BigButton';
 import SkipButton from '../components/SkipButton';
 import SubmitButton from '../components/SubmitButton'
 import Language from '../containers/Language';
 import MultiSelect from '../containers/MultiSelect';
-import { FormattedMessage } from 'react-intl';
-import { string, func } from 'prop-types'
+import {
+    FormattedMessage,
+} from 'react-intl';
 
 import {
     SELECT,
@@ -57,15 +68,15 @@ class QuestionPage extends React.Component {
     }
 
     static propTypes = {
-        question: PropTypes.object.isRequired,
-        questionChoices: PropTypes.arrayOf(PropTypes.object).isRequired,
-        onChoiceClick: PropTypes.func.isRequired,
-        questionType: PropTypes.string.isRequired,
-        onSubmitFreeText: PropTypes.func.isRequired,
-        questionPos: PropTypes.number.isRequired,
-        error: PropTypes.bool,
-        moveToNextQuestion: PropTypes.func.isRequired,
-        currentIsRequired: PropTypes.bool.isRequired,
+        question: object.isRequired,
+        questionChoices: arrayOf(object).isRequired,
+        onChoiceClick: func.isRequired,
+        questionType: string.isRequired,
+        onSubmitFreeText: func.isRequired,
+        questionPos: number.isRequired,
+        error: bool,
+        moveToNextQuestion: func.isRequired,
+        currentIsRequired: bool.isRequired,
         text: string.isRequired,
         addAnswer: func.isRequired,
         showFieldRequired: func.isRequired,
